% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cod4_conf_sam}
\alias{cod4_conf_sam}
\title{North Sea cod SAM configuration file}
\format{An object of class \code{list} of length 20.}
\source{
\url{https://www.stockassessment.org}
}
\usage{
cod4_conf_sam
}
\description{
Configuration file for the SAM stockassessment for North Sea cod as used by
WGNSSK 2017.
}
\keyword{datasets}
