% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cod4_idx}
\alias{cod4_idx}
\title{North Sea cod indices}
\format{An object of class \code{FLIndices} with elements:
\describe{
  \item{IBTS_Q1_gam}{An object of class \code{FLIndex}}
  \item{IBTS_Q3_gam}{An object of class \code{FLIndex}}
}}
\source{
\url{https://www.stockassessment.org}
}
\usage{
cod4_idx
}
\description{
Survey indices for North Sea cod. The time
series covers the years 1963-2016. Two surveys are included: Q1 IBTS and
Q3 IBTS. The survey values are the values as used by WGNSSk 2017 and not the
raw values.
}
\keyword{datasets}
