% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cod4_stk}
\alias{cod4_stk}
\title{North Sea cod}
\format{An object of class \code{FLStock}}
\source{
\url{https://www.stockassessment.org}
}
\usage{
cod4_stk
}
\description{
Assessment input data for North Sea cod as used by ICES WGNSSK 2017. The time
series covers the years 1963-2017, with catches up to and including 2016.
}
\keyword{datasets}
