% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cod4_stk_2019}
\alias{cod4_stk_2019}
\title{North Sea cod}
\format{An object of class \code{FLStock}}
\source{
\url{https://www.stockassessment.org}
}
\usage{
cod4_stk_2019
}
\description{
Assessment input data for North Sea cod as used by ICES WGNSSK 2019. The time
series covers the years 1963-2019, with catches up to and including 2018.
}
\keyword{datasets}
