% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Condition_OM_stock.R
\name{create_FLIndex}
\alias{create_FLIndex}
\title{Function to create FLindex of the OM for MSE simulation}
\usage{
create_FLIndex(idx, stk, stk0, it)
}
\arguments{
\item{idx}{object of class FLindex}

\item{stk}{object of class FLStock}

\item{stk0}{median of stk iterations created in \link[FLfse]{create_FLStock}}

\item{it}{iterations}
}
\value{
FLindex
}
\description{
Function to create FLindex of the OM for MSE simulation
}
\details{
creates the FLindex, script based on
http://www.flr-project.org/doc/An_introduction_to_MSE_using_FLR.html
}
