% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_om.R
\name{create_FLStock}
\alias{create_FLStock}
\title{Function to create several iterations of FLStock}
\usage{
create_FLStock(stk, idx, it, qmod = NULL, fmod = NULL, mcsave = 100,
  seed.nb = 321)
}
\arguments{
\item{stk}{object of class FLStock}

\item{idx}{object of class FLIndices}

\item{it}{iterations}

\item{qmod}{catchability submodel}

\item{fmod}{fishing mortality submodel}

\item{mcsave}{mcmc parameter, default 100}

\item{seed.nb}{set seed number}
}
\value{
a list of "stk" and "stk0", i.e. the median of all iterations
}
\description{
Function to create several iterations of FLStock
}
\details{
creates \code{it} iterations of the FLStock \code{stk} using MCMC,
calls a4aSCA to run the assessment on existing FLStock \code{stk} and 
FLIndices \code{idx} objects
script based on
http://www.flr-project.org/doc/An_introduction_to_MSE_using_FLR.html
}
