% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_om_biol.R
\name{create_FLStock_biol}
\alias{create_FLStock_biol}
\title{Function to create FLStock of the OM for MSE simulation
from biological data using existing dataset of WKLife or
just providing own parameters}
\usage{
create_FLStock_biol(latin.name = NULL, common.name = NULL, region = NULL,
  biol.params = NULL, h = 0.75, it = 5, seed.nb = 321)
}
\arguments{
\item{latin.name}{vector of latin names of species to be selected from WKLife dataset
(if given, no need for \code{common.name})}

\item{common.name}{vector of common names of species to be selected from WKLife dataset
(if given, no need for \code{latin.name})}

\item{region}{stock area from WKLife dataset}

\item{biol.params}{dataframe of biological params if not using from WKLife stocks 
(i.e. if not using the 3 above arguments). No need to provide all variables as FLife is used to 
complete the set}

\item{h}{steepness default as 0.75 (can ba a vecor of same length as number of species or stocks)}

\item{it}{number of iterations, default is 5}

\item{seed.nb}{set seed number}
}
\value{
a list of "sr", "brp", "stk"
}
\description{
Function to create FLStock of the OM for MSE simulation
from biological data using existing dataset of WKLife or
just providing own parameters
}
\details{
creates the FLStock, script based on
https://github.com/shfischer/wklifeVII/blob/master/R/OM1.R
}
\examples{
\dontrun{

# Example using given life history parameter(s)
test1 <- create_FLStock_biol(biol.params=data.frame(linf=100))
stock.n(test1[[1]]$stk)
plot(test1[[1]]$stk)

# Example using WKLife dataset
test2 <- create_FLStock_biol(latin.name="Clupea harengus")
stock.n(test2[[1]]$stk)
plot(test2[[1]]$stk)
}
}
