% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{FLR_SAM}
\alias{FLR_SAM}
\title{Run SAM with FLR objects.}
\usage{
FLR_SAM(stk, idx, conf = NULL)
}
\arguments{
\item{stk}{Object of class \linkS4class{FLStock} with stock and fishery data.}

\item{idx}{Object of class \linkS4class{FLIndices} or \linkS4class{FLIndex}
object with survey index time series.}

\item{conf}{Optional configurations passed to SAM. Should be a list.}
}
\value{
An object of class \code{spictcls} with the model results.
}
\description{
This function runs a SAM assessment using FLR objects as input. Stock and
fishery data is extracted from an object of class \code{FLStock}, survey
indices from \code{FLIndex} or \code{FLIndices}. Additional model
configurations can be passed to SAM.
}
\details{
Stock and fishery data is extracted from \code{stk}, survey indices from
\code{idx}.
Stock data used are:
\itemize{
  \item catch numbers at age (\code{catch.n} slot from \code{stk})
  \item maturity ogive (\code{mat} slot from \code{stk})
  \item stock weight at age (\code{stock.wt} slot from \code{stk})
  \item catch weight at age (\code{catch.wt} slot from \code{stk})
  \item discards weigth at age (\code{discards.wt} slot from \code{stk})
  \item landings weight at age (\code{landings.wt} slot from \code{stk})
  \item natural mortality at age (\code{m} slot from \code{stk})
  \item proportion of fishing mortality before spawning at age
    (\code{harvest.spwn} slot from \code{stk})
   \item proportion of natural mortality before spawning at age
     (\code{catch.n} slot from \code{stk})
   \item landing fraction at age (calculated as landings numbers at age
     divided by the catch numbers at age)
}

Trailing years without data in \code{stk} are removed.

Survey indices are extracted from \code{idx}, using the \code{index}
slot(s).
SSB indices can be used. To define an index as SSB index, the first (age)
dimension of the \code{index} slot of \code{idx} has to be of length 1 and
the name of this dimension can be either missing (\code{NA}), non-numeric
or \code{-1}.

Additional configurations can be passed as a list to SAM with the
\code{conf} argument. \code{FLR_SAM} first generates a default model
configuration with \code{stockassessment}'s \code{setup.sam.data}. If
additional configurations are passed to \code{FLR_SAM}, they will replace
the default configuration. For details about possible configurations and
format see '\code{help("defcon", package = "stockassessment")} and
\url{https://github.com/fishfollower/SAM}.
}
\section{Warning}{

This methods requires the \code{stockassessment} package and all its
dependencies to be installed. For details how to obtain \code{spict}, see
\url{https://github.com/fishfollower/SAM/}.
}

\examples{
# load example data

}
