% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spict.R
\name{FLR_SPiCT}
\alias{FLR_SPiCT}
\title{Run SPiCT with FLR objects.}
\usage{
FLR_SPiCT(stk, idx, conf = NULL)
}
\arguments{
\item{stk}{Object of class \linkS4class{FLStock} with catch time series.}

\item{idx}{Object of class \linkS4class{FLIndices} or \linkS4class{FLIndex}
object with survey index time series.}

\item{conf}{Optional configurations passed to SPiCT. Should be a list.}
}
\value{
An object of class \code{spictcls} with the model results.
}
\description{
This function extracts the catch from an \code{FLStock} object and the survey
indices/index from a \code{FLIndices} or \code{FLIndex} object and runs a
SPiCT (Surplus Production in Continuous Time) stock assessment.
}
\details{
The catch time series is obtained from the \code{catch} slot of \code{stk.}

The survey index/indices are obtained from the \code{index} slot(s) of
\code{idx}. If the slot(s) contains an age age structure, the sum over all
ages is used.

Additional configurations can be passed as a list to SPiCT with the
\code{conf} argument. They are passed directly to SPiCT
(\code{\link[spict]{fit.spict}}) without checking. Any configurations accepted by
(\code{\link[spict]{fit.spict}}) can be used.
}
\section{Warning}{

This methods requires the \code{spict} package and all its dependencies to be
  installed. For details how to obtain \code{spict}, see
  \url{https://github.com/mawp/spict/}.
}

\examples{
# fit SPiCT to Irish Sea plaice
fit <- FLR_SPiCT(stk = ple7a_stk, idx = ple7a_idx)
fit

# pass additional configuration, set time step to 1 per year
conf <- list(dteuler = 1)
fit <- FLR_SPiCT(stk = ple7a_stk, idx = ple7a_idx, conf = conf)
fit

}
