% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getCtrl}
\alias{getCtrl}
\title{Define control object for \link[FLash]{fwd}}
\usage{
getCtrl(values, quantity, years)
}
\arguments{
\item{values}{}

\item{quantity}{}

\item{years}{}
}
\value{
ctrl for fwd
}
\description{
Define control object for \link[FLash]{fwd}
}
\details{
The fwd method from FLash needs a control object, which is set by this function.
}
