% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\docType{methods}
\name{getpars}
\alias{getpars}
\alias{getpars,sam-method}
\alias{getpars,sam_list-method}
\title{Get parameter estimates from SAM model fit.}
\usage{
getpars(fit)

\S4method{getpars}{sam}(fit)

\S4method{getpars}{sam_list}(fit)
}
\arguments{
\item{fit}{A single SAM model fit of class \code{sam} or a list of fits of
class \code{sam_lst}.}
}
\value{
A list with the model parameters of the SAM or a list of them in case
of several supplied models
}
\description{
This function extracts the parameter estimates from a SAM model fit. These
are useful e.g. as initial values in subsequent model fits and can improve
model convergence/computing time.
}
\examples{
### fit SAM to North Sea cod
fit <- FLR_SAM(stk = cod4_stk, idx = cod4_idx, conf = cod4_conf_sam)

### extract parameters
pars <- getpars(fit)

### use them as starting values
fit2 <- FLR_SAM(stk = cod4_stk, idx = cod4_idx, conf = cod4_conf_sam,
                par_ini = pars)
}
