% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{had4_conf_sam}
\alias{had4_conf_sam}
\title{North Sea haddock SAM configuration file}
\format{An object of class \code{list} of length 20.}
\source{
\url{https://www.stockassessment.org}
}
\usage{
had4_conf_sam
}
\description{
Configuration file for the SAM stockassessment for North Sea haddock. The
configuration is simply the default configuration derived from SAM.
}
\keyword{datasets}
