% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{had4_idx}
\alias{had4_idx}
\title{North Sea haddock indices}
\format{An object of class \code{FLIndices} with elements:
\describe{
  \item{IBTS_Q1_gam}{An object of class \code{FLIndex}}
  \item{IBTS_Q3_gam}{An object of class \code{FLIndex}}
}}
\source{
\url{https://www.stockassessment.org}
}
\usage{
had4_idx
}
\description{
Survey indices for North Sea cod. Two surveys are included: Q1 IBTS and
Q3 IBTS. The time series covers the years 1967-2018 for Q1 and 1991-2017 for
Q3.The survey values are the values as used by WGNSSk 2018 and not the
raw values.
}
\keyword{datasets}
