% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{had4_stk}
\alias{had4_stk}
\alias{had4_idx}
\alias{had4_conf_sam}
\title{North Sea haddock}
\format{
An object of class \code{FLStock} of dimension 9 x 54 x 1 x 1 x 1 x 1.

An object of class \code{FLIndices} of length 2.

An object of class \code{list} of length 20.
}
\source{
\url{https://www.stockassessment.org}
}
\usage{
had4_stk

had4_idx

had4_conf_sam
}
\description{
Haddock (\emph{Melanogrammus aeglefinus}) in Subarea 4, Division 6.a, and
Subdivision 20 (North Sea, West of Scotland, Skagerrak).
Assessment input data for North Sea haddock as used by ICES WGNSSK 2018.
Includes the stock, the indices and the SAM configuration. See the example
below for how to run the assessment.
}
\examples{
# Replicate the 2018 assessment:
fit <- FLR_SAM(stk = had4_stk,
               idx = had4_idx,
               conf = had4_stk)

}
\keyword{datasets}
