% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{her4_conf_sam}
\alias{her4_conf_sam}
\title{North Sea herring SAM configuration file}
\format{An object of class \code{list} of length 21.}
\source{
\url{https://github.com/ices-eg/wg_HAWG/tree/master/NSAS}
}
\usage{
her4_conf_sam
}
\description{
Herring (Clupea harengus) in Subarea 4 and divisions 3.a and 7.d, autumn
spawners (North Sea, Skagerrak and Kattegat, eastern English Channel).
Configuration file for the SAM stockassessment for North Sea herring as used
by ICES HAWG 2019.
}
\keyword{datasets}
