% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{her4_idx}
\alias{her4_idx}
\title{North Sea herring indices}
\format{An object of class \code{FLIndices} with elements:
\describe{
  \item{HERAS}{An object of class \code{FLIndex}}
  \item{IBTS-Q1}{An object of class \code{FLIndex}}
  \item{IBTS0}{An object of class \code{FLIndex}}
  \item{IBTS-Q3}{An object of class \code{FLIndex}}
  \item{LAI-ORSH}{An object of class \code{FLIndex}}
  \item{LAI-BUN}{An object of class \code{FLIndex}}
  \item{LAI-CNS}{An object of class \code{FLIndex}}
  \item{LAI-SNS}{An object of class \code{FLIndex}}
}}
\source{
\url{https://github.com/ices-eg/wg_HAWG/tree/master/NSAS}
}
\usage{
her4_idx
}
\description{
Herring (Clupea harengus) in Subarea 4 and divisions 3.a and 7.d, autumn
spawners (North Sea, Skagerrak and Kattegat, eastern English Channel).
Survey indices for North Sea herring. Includes data for eight surveys:
HERAS, IBTS-Q1, IBTS0", IBTS-Q3, LAI-ORSH, LAI-BUN, LAI-CNS and LAI-SNS.
The survey values are the values as used by ICES HAWG 2019 and not the
raw values.
}
\keyword{datasets}
