% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{her4_stk}
\alias{her4_stk}
\title{North Sea herring}
\format{An object of class \code{FLStock}}
\source{
\url{https://github.com/ices-eg/wg_HAWG/tree/master/NSAS}
}
\usage{
her4_stk
}
\description{
Herring (Clupea harengus) in Subarea 4 and divisions 3.a and 7.d, autumn
spawners (North Sea, Skagerrak and Kattegat, eastern English Channel).
Assessment input data for North Sea herring as used by ICES HAWG 2019.
The time series covers the years 1947-2019, with catches up to and
including 2018.
}
\keyword{datasets}
