% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mac.27.nea_stk_2019}
\alias{mac.27.nea_stk_2019}
\alias{mac.27.nea_idx_2019}
\alias{mac.27.nea_conf_2019}
\title{Northeast Atlantic mackerel 2019}
\format{
An object of class \code{FLStock} of dimension 13 x 40 x 1 x 1 x 1 x 1.

An object of class \code{FLIndices} of length 3.

An object of class \code{list} of length 23.
}
\source{
\url{https://www.stockassessment.org}
}
\usage{
mac.27.nea_stk_2019

mac.27.nea_idx_2019

mac.27.nea_conf_2019
}
\description{
Mackerel  (\emph{Scomber  scombrus}) in subareas 1-8 and 14, and in
Division 9.a (the Northeast Atlantic and adjacent waters).
Assessment input data for Northeast Atlantic mackerel as used by ICES WGWIDE
2019. Includes the stock, the indices and the SAM configuration.
}
\details{
\code{mac.27.nea_stk_2019} includes tagging data, stored as an attribute in
\code{attr(catch.n(mac.27.nea_stk_2019), "recap")}. An additional tagging
configuration is stored in
\code{attr(catch.n(mac.27.nea_stk_2019), "recap_conf")}.
Also, weights for the catch numbers are supplied as an attribute in
\code{attr(catch.n(mac.27.nea_stk_2019), "weight")}. All additional data is
passed on automatically to the SAM assessment.
}
\examples{
# The 2019 WGWIDE Northeast Atlantic mackerel assessment can be reproduced
# with:
fit <- FLR_SAM(stk = mac.27.nea_stk_2019,
               idx = mac.27.nea_idx_2019,
               conf = mac.27.nea_conf_2019)

}
\keyword{datasets}
