% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{observation_error_proj}
\alias{observation_error_proj}
\title{Function to include error in the projected index}
\usage{
observation_error_proj(stk, idx, assessmentYear, dataYears)
}
\arguments{
\item{stk}{object of class FLStock}

\item{idx}{object of class FLindex}

\item{assessmentYear}{latest year of the assessment}

\item{dataYears}{position vector of years with data}
}
\value{
a list of stock, new index and index of the OM
}
\description{
Function to include error in the projected index
}
