% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_sr_setup.R
\name{om_sr_model}
\alias{om_sr_model}
\title{Fit an SR model to FLStock and create residuals for MSE projection}
\usage{
om_sr_model(stk = NULL, srbh.res = NULL, sr_model = "bevholt",
  method = "L-BFGS-B", it, iy, fy, seed.nb = 321)
}
\arguments{
\item{stk}{FLStock - if NULL the \code{sr_model} will be used to create 
projected residuals}

\item{srbh.res}{residuals of FLSR - default is NULL but if given, 
no need for \code{stk}}

\item{sr_model}{default is \code{"bevholt"}}

\item{method}{default is \code{"L-BFGS-B"}, not fit for other methods yet}

\item{it}{number of iterations}

\item{iy}{intermediate year (first year of projection)}

\item{fy}{final year of projection}

\item{seed.nb}{set seed number}
}
\value{
a list of "srbh" (FLSR), "srbh0", i.e. the median of all iterations, and
"srbh.res", i.e. recruitment residuals for projection period if \code{stk} given,
or only "srbh.res" if \code{srbh.res} given
}
\description{
Fit an SR model to FLStock and create residuals for MSE projection
}
