% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ple7a_conf_sam}
\alias{ple7a_conf_sam}
\title{Irish Sea plaice SAM configuration file}
\format{An object of class \code{list} of length 3.}
\usage{
ple7a_conf_sam
}
\description{
Configuration file for the SAM stockassessment for Irish Sea plaice as used
by WGCSE 2017.
}
\keyword{datasets}
