% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ple7a_idx}
\alias{ple7a_idx}
\title{Irish Sea plaice indices}
\format{An object of class \code{FLIndices} with elements:
\describe{
  \item{UK BT SURVEY (Sept) - all stations Rcode}{An object of class
    \code{FLIndex}}
  \item{DARDS}{An object of class \code{FLIndex}}
  \item{DARDA}{An object of class \code{FLIndex}}
}}
\usage{
ple7a_idx
}
\description{
Survey indices for Irish Sea plaice as used by WGCSE 2017. Three survey
indices are included: "UK BT SURVEY (Sept) - all stations Rcode", "DARDS"
and "DARDA".
}
\keyword{datasets}
