% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_points.R
\name{ref_pts}
\alias{ref_pts}
\title{Simple list of reference points}
\usage{
ref_pts(stk, srbh)
}
\arguments{
\item{stk}{FLStock (median value if several iterations)}

\item{srbh}{FLSR (median value if several iterations)}
}
\value{
list of reference points
}
\description{
needs to be tidied up for options
}
