% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{SAM2FLStock}
\alias{SAM2FLStock}
\alias{SAM2FLStock,sam,missing-method}
\alias{SAM2FLStock,sam_list,missing-method}
\alias{SAM2FLStock,sam,FLStock-method}
\alias{SAM2FLStock,sam_list,FLStock-method}
\title{Coerce SAM output into \code{FLStock} object}
\usage{
SAM2FLStock(
  object,
  stk,
  uncertainty = FALSE,
  conf_level = 95,
  stock_only = FALSE,
  catch_estimate = FALSE,
  correct_catch = FALSE,
  mat_est = FALSE,
  stock.wt_est = FALSE,
  catch.wt_est = FALSE,
  m_est = FALSE,
  spinoutyear = FALSE
)

\S4method{SAM2FLStock}{sam,missing}(
  object,
  stk,
  uncertainty = FALSE,
  conf_level = 95,
  stock_only = FALSE,
  catch_estimate = FALSE,
  correct_catch = FALSE,
  mat_est = FALSE,
  stock.wt_est = FALSE,
  catch.wt_est = FALSE,
  m_est = FALSE,
  spinoutyear = FALSE
)

\S4method{SAM2FLStock}{sam_list,missing}(
  object,
  stk,
  uncertainty = FALSE,
  conf_level = 95,
  stock_only = FALSE,
  catch_estimate = FALSE,
  correct_catch = FALSE,
  mat_est = FALSE,
  stock.wt_est = FALSE,
  catch.wt_est = FALSE,
  m_est = FALSE,
  spinoutyear = FALSE
)

\S4method{SAM2FLStock}{sam,FLStock}(
  object,
  stk,
  uncertainty = FALSE,
  conf_level = 95,
  stock_only = FALSE,
  catch_estimate = FALSE,
  correct_catch = FALSE,
  mat_est = FALSE,
  stock.wt_est = FALSE,
  catch.wt_est = FALSE,
  m_est = FALSE,
  spinoutyear = FALSE
)

\S4method{SAM2FLStock}{sam_list,FLStock}(
  object,
  stk,
  uncertainty = FALSE,
  conf_level = 95,
  stock_only = FALSE,
  catch_estimate = FALSE,
  correct_catch = FALSE,
  mat_est = FALSE,
  stock.wt_est = FALSE,
  catch.wt_est = FALSE,
  m_est = FALSE,
  spinoutyear = FALSE
)
}
\arguments{
\item{object}{Object of class \code{sam} with the results from a
SAM stock assessment run. Alternatively, object of class \code{sam_list},
i.e. a list of \code{sam} objects.}

\item{stk}{Optional. Object of class \linkS4class{FLStock}, to which the
assessment results are added.}

\item{uncertainty}{If set to \code{TRUE}, the estimated uncertainty from
SAM will be added as attribute.}

\item{conf_level}{Confidence level used when uncertainty is returned.
Defaults to 95 (percent).}

\item{stock_only}{Logical. If set to \code{TRUE}, catch data (numbers,
weights) are ignored and only stock data (numbers, SSB, etc.) are
returned}

\item{catch_estimate}{Logical, return the catch estimated by SAM instead of the model input?}

\item{correct_catch}{Logical, correct catch with
catch multiplier estimated by SAM?}

\item{mat_est}{Logical, return SAM estimates for maturity?}

\item{stock.wt_est}{Logical, return SAM estimates for stock weights?}

\item{catch.wt_est}{Logical, return SAM estimates for catch weights?}

\item{m_est}{Logical, return SAM estimates for natural mortality?}

\item{spinoutyear}{Logical, return SAM estimates of biological estimates beyond last data year?}
}
\value{
An object of class \code{FLStock}.
}
\description{
This function takes the output from running the SAM stockassessment and
converts them into an \code{FLStock} object.
}
\details{
\code{SAM2FLStock} returns both the input data used for running SAM (e.g. catch) and the model estimates (stock numbers and fishing mortality). By default, the returned catch is the input catch provided to SAM. However, the catch as estimated by SAM can be returned by setting \code{catch_estimate = TRUE}. Also, estimates of biological data (stock weights, catch weights, natural mortality, maturity) can be returned if requested and available from the model fit. Setting \code{uncertainty = TRUE} returns confidence intervals for catch, stock numbers and fishing mortality, saved as attributes in the corresponding slots of the \code{FLStock} output.

If an \code{FLStock} is provided as \code{stk} argument, then this is used as template. If the dimensions (years, iterations) differ between the SAM results and the provided stock template, the returned \code{FLStock} is expanded.

The \code{object} argument can either be a single SAM model fit or a list of SAM model fits (defined as class \code{sam_list}). If a list is provided, the output is an \code{FLStock} object where the different iterations correspond to the individual model fits.

The function can handle SAM model fits with multiple fleets. In the returned
FLStock, the fleets are combined into a single fleet. Some functionality
(e.g. uncertainty bounds) might not work for multiple fleets.
}
\examples{
# fit SAM to North Sea cod
fit <- FLR_SAM(stk = cod4_stk, idx = cod4_idx, conf = cod4_conf_sam)

# coerce the output into FLStock
stk <- SAM2FLStock(fit)

# get catch estimates from model
stk <- SAM2FLStock(fit, catch_estimate = TRUE)

\dontrun{
# use multi-fleet SAM model for western Baltic spring-spawning herring and
# load model fit from stockassessment.org
fit <- stockassessment::fitfromweb("WBSS_HAWG_2021")
stk <- SAM2FLStock(fit)
}

}
