% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{whg4_conf_sam}
\alias{whg4_conf_sam}
\title{North Sea whiting SAM configuration file}
\format{An object of class \code{list} of length 20.}
\source{
\url{https://www.stockassessment.org}
}
\usage{
whg4_conf_sam
}
\description{
Configuration file for the SAM stockassessment for North Sea whiting as used
 by WGNSSK 2018.
}
\keyword{datasets}
