% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{whg4_idx}
\alias{whg4_idx}
\title{North Sea whiting indices}
\format{An object of class \code{FLIndices} with elements:
\describe{
  \item{IBTS-Q1}{An object of class \code{FLIndex}}
  \item{IBTS-Q3}{An object of class \code{FLIndex}}
}}
\source{
\url{https://www.stockassessment.org}
}
\usage{
whg4_idx
}
\description{
Survey indices for North Sea whiting The time
series covers the years 1983-2018. Two surveys are included: Q1 IBTS and
Q3 IBTS. The survey values are the values as used by WGNSSk 2018 and not the
raw values.
}
\keyword{datasets}
