% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_model.R
\name{create_FLIndices}
\alias{create_FLIndices}
\title{Function to create FLIndices of the OM for MSE simulation}
\usage{
create_FLIndices(idx, stk, stk0, it)
}
\arguments{
\item{idx}{object of class FLIndices}

\item{stk}{object of class FLStock}

\item{stk0}{median of stk iterations created in \link[FLfse]{create_FLStock}}

\item{it}{iterations}
}
\value{
FLIndices
}
\description{
Function to create FLIndices of the OM for MSE simulation
}
\details{
creates the FLIndices, script based on
http://www.flr-project.org/doc/An_introduction_to_MSE_using_FLR.html
}
