% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_om.R
\name{create_FLStock}
\alias{create_FLStock}
\title{Function to create FLStock of the OM for MSE simulation}
\usage{
create_FLStock(stk, idx, it, qmod = NULL, fmod = NULL, mcsave = 100)
}
\arguments{
\item{stk}{object of class FLStock}

\item{idx}{object of class FLIndices}

\item{it}{iterations}

\item{qmod}{catchability submodel}

\item{fmod}{fishing mortality submodel
default ~te(replace(age, age>9,9), year, k=c(6,8))}

\item{mcsave}{mcmc parameter, default 100}

\item{ny}{number of years to project from initial year (iy = maxyear of stk + 1)}

\item{nsqy}{number of years to compute status quo metrics, default set at 3}

\item{sr}{stock-recruit model, default is bevholt}
}
\value{
a list of "stk", "dy", "stk0", "srbh","srbh0","srbh.res","brp","Fmsy","msy","Bmsy","Bpa","Blim"
}
\description{
Function to create FLStock of the OM for MSE simulation
}
\details{
creates the FLStock, script based on
http://www.flr-project.org/doc/An_introduction_to_MSE_using_FLR.html
It uses a4aSCA to run the assessment on existing FLStock and FLIndices objects
}
