% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\docType{methods}
\name{FLR_SAM}
\alias{FLR_SAM}
\alias{FLR_SAM,FLStock,FLIndices-method}
\alias{FLR_SAM,FLStock,FLIndex-method}
\title{Run SAM with FLR objects.}
\usage{
FLR_SAM(stk, idx, conf = NULL, DoParallel = FALSE, ...)

\S4method{FLR_SAM}{FLStock,FLIndices}(stk, idx, conf = NULL,
  DoParallel = FALSE, ...)

\S4method{FLR_SAM}{FLStock,FLIndex}(stk, idx, conf = NULL,
  DoParallel = FALSE, ...)
}
\arguments{
\item{stk}{Object of class \linkS4class{FLStock} with stock and fishery data.}

\item{idx}{Object of class \linkS4class{FLIndices} or \linkS4class{FLIndex}
object with survey index time series.}

\item{conf}{Optional configurations passed to SAM. Defaults to \code{NULL}.
If provided, should be a list.}

\item{DoParallel}{Optional, defaults to \code{FALSE}. If set to \code{TRUE},
will perform iterations of stock in parallel. See Details below for
description.}

\item{...}{Additional arguments passed to \code{sam.fit()}, e.g.
\code{newtonsteps}}
}
\value{
An object of class \code{sam} (for single iteration) or
  \code{sam_list} (list of \code{sam} objects for multiple iterations) with
  the model results.
}
\description{
This function runs a SAM assessment using FLR objects as input. Stock and
fishery data is extracted from an object of class \code{FLStock}, survey
indices from \code{FLIndex} or \code{FLIndices}. Additional model
configurations can be passed to SAM.
}
\details{
Stock and fishery data is extracted from \code{stk}, survey indices from
\code{idx}.
Stock data used are:
\itemize{
  \item catch numbers at age (\code{catch.n} slot from \code{stk})
  \item maturity ogive (\code{mat} slot from \code{stk})
  \item stock weight at age (\code{stock.wt} slot from \code{stk})
  \item catch weight at age (\code{catch.wt} slot from \code{stk})
  \item discards weigth at age (\code{discards.wt} slot from \code{stk})
  \item landings weight at age (\code{landings.wt} slot from \code{stk})
  \item natural mortality at age (\code{m} slot from \code{stk})
  \item proportion of fishing mortality before spawning at age
    (\code{harvest.spwn} slot from \code{stk})
   \item proportion of natural mortality before spawning at age
     (\code{catch.n} slot from \code{stk})
   \item landing fraction at age (calculated as landings numbers at age
     divided by the catch numbers at age)
}

Trailing years without data in \code{stk} are removed.

Survey indices are extracted from \code{idx}, using the \code{index}
slot(s).
SSB indices can be used. To define an index as SSB index, the first (age)
dimension of the \code{index} slot of \code{idx} has to be of length 1 and
the name of this dimension can be either missing (\code{NA}), non-numeric
or \code{-1}.

Additional configurations can be passed as a list to SAM with the
\code{conf} argument. \code{FLR_SAM} first generates a default model
configuration with \code{stockassessment}'s \code{setup.sam.data}. If
additional configurations are passed to \code{FLR_SAM}, they will replace
the default configuration. For details about possible configurations and
format see '\code{help("defcon", package = "stockassessment")} and
\url{https://github.com/fishfollower/SAM}.

The function can handle input objects with multiple iterations (\code{iter}
dimension in \code{stk} and \code{idx}). If multiple iterations are provided
for \code{stk} but not for \code{idx}, \code{idx} will be inflated, and vice
 versa.
If the assessment fails for some iterations, the error messages are return
for these iterations.
If argument \code{DoParallel} is set to \code{TRUE}, the individual
iterations are processed in parallel. This uses parallel computing provided
by the package \code{DoParallel}. The parallel workers need to be set up
and registered before calling the function. See \code{?DoParallel}.
}
\section{Warning}{

This methods requires the \code{stockassessment} package and all its
dependencies to be installed. For details how to obtain
\code{stockassessment}, see \url{https://github.com/fishfollower/SAM/}.
}

\examples{
# fit SAM to North Sea cod
fit <- FLR_SAM(stk = cod4_stk, idx = cod4_idx)

# use WGNSSK 2017 configuration
fit <- FLR_SAM(stk = cod4_stk, idx = cod4_idx, conf = cod4_conf_sam)

# fit SAM to Irish Sea plaice
fit <- FLR_SAM(stk = ple7a_stk, idx = ple7a_idx, conf = ple7a_conf_sam)

}
