% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{had4_stk}
\alias{had4_stk}
\title{North Sea haddock}
\format{An object of class \code{FLStock}}
\source{
\url{https://www.stockassessment.org}
}
\usage{
had4_stk
}
\description{
Assessment input data for North Sea haddock as used by ICES WGNSSK 2018.
The time series covers the years 1963-2018, with catches up to and
including 2017.
}
\keyword{datasets}
