% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse_base.R
\name{mse_base}
\alias{mse_base}
\title{Run the full MSE}
\usage{
mse_base(stk, idx, it, ny, assessment = "sam")
}
\arguments{
\item{stk}{object of class FLStock}

\item{idx}{object of class FLindex}

\item{it}{}

\item{ny}{number of years to project from initial year (iy = maxyear of stk + 1)}

\item{assessment}{assessment model to be used, default is sam (will also be set up for spict)}
}
\value{
mse runs
}
\description{
Run the full MSE
}
\examples{
\dontrun{
stk = ple4
idx = FLIndices(idx=ple4.index)
}
}
