% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse_function.R
\name{mse_fn}
\alias{mse_fn}
\title{Run the MSE loop}
\usage{
mse_fn(stk, idx, y0, iy, dy, fy, nsqy, srbh, srbh.res, assessment = "sam",
  Bpa, Fmsy, seed.nb = 321)
}
\arguments{
\item{stk}{object of class FLStock}

\item{idx}{object of class FLindex}

\item{y0}{intial data year}

\item{iy}{intermediate year}

\item{dy}{final data year of original FLStock}

\item{fy}{total number of years}

\item{nsqy}{number of years to compute status quo metrics, default set at 3}

\item{srbh}{stock-recruit model}

\item{srbh.res}{stock-recruit model residuals}

\item{assessment}{assessment model to be used, default is sam (will also be set up for spict)}

\item{Bpa}{reference point}

\item{Fmsy}{reference points}

\item{seed.nb}{sets seed}
}
\value{
mse runs
}
\description{
Run the MSE loop
}
