% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_sr_setup.R
\name{om_sr_model}
\alias{om_sr_model}
\title{Fit an SR model to FLStock and create residuals for MSE projection}
\usage{
om_sr_model(stk, sr_model = "bevholt", method = "L-BFGS-B", it, iy, fy,
  seed.nb = 321)
}
\arguments{
\item{stk}{FLStock}

\item{sr_model}{default is "bevholt"}

\item{method}{default is "L-BFGS-B", not fit for other methods yet}

\item{it}{number of iterations}

\item{iy}{intermediate year (first year of projection)}

\item{fy}{final year of projection}

\item{seed.nb}{set seed number}
}
\value{
list of median FLSotck, FLSR for FLStock and median FLStock for reference points and
residuals for projection
}
\description{
Fit an SR model to FLStock and create residuals for MSE projection
}
