% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ple7a_stk}
\alias{ple7a_stk}
\title{Irish Sea plaice}
\format{An object of class \code{FLStock}}
\usage{
ple7a_stk
}
\description{
Assessment input data for Irish Sea plaice as used by ICES WGCSE 2017.
The time series covers the years 1963-2016.
}
\keyword{datasets}
