% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_points.R
\name{ref_pts}
\alias{ref_pts}
\title{Simple list of reference points}
\usage{
ref_pts(stk = NULL, srbh = NULL, brps = NULL)
}
\arguments{
\item{stk}{FLStock (provide median value if several iterations)}

\item{srbh}{FLSR (provide median value if several iterations)}

\item{brps}{FLBRP if already existing (no need for other \code{stk}
or \code{srbh} in that case)}
}
\value{
list of reference points
}
\description{
needs to be tidied up for options
}
