% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse_wrapper.R
\name{run_mse}
\alias{run_mse}
\title{Master function to run MSE simulations}
\usage{
run_mse(scenarios, parallel = FALSE)
}
\arguments{
\item{scenarios}{list of scenarios to run}

\item{parallel}{if TRUE, will run scenarios in parallel. Default is F}
}
\value{
output of mse_base
}
\description{
This is the main high-level wrapper function for running \pkg{FLfse}
simulations.
}
\examples{
\dontrun{
simple_scenario1 <- list(scens = list(id = "C.hareng"), 
                         stock = list(latin.name="Clupea harengus"),
                         ctrl.sims = list(it = 2, ny = 3))
simple_scenario2 <- list(scens = list(id = "Linf=100"),
                         stock = list(biol.params=data.frame(linf=100)),
                         ctrl.sims = list(it = 2, ny = 3))

test1 <- run_mse(scenarios=list(simple_scenario1, simple_scenario2), parallel = F)
test2 <- run_mse(scenarios=list(simple_scenario1, simple_scenario2), parallel = T)
}

}
