% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{SAM2FLStock}
\alias{SAM2FLStock}
\alias{SAM2FLStock,sam,missing-method}
\alias{SAM2FLStock,sam_list,missing-method}
\alias{SAM2FLStock,sam,FLStock-method}
\alias{SAM2FLStock,sam_list,FLStock-method}
\title{Coerce SAM output into \code{FLStock} object}
\usage{
SAM2FLStock(
  object,
  stk,
  uncertainty = FALSE,
  conf_level = 95,
  catch_estimate = FALSE,
  correct_catch = FALSE
)

\S4method{SAM2FLStock}{sam,missing}(
  object,
  stk,
  uncertainty = FALSE,
  conf_level = 95,
  catch_estimate = FALSE,
  correct_catch = FALSE
)

\S4method{SAM2FLStock}{sam_list,missing}(
  object,
  stk,
  uncertainty = FALSE,
  conf_level = 95,
  catch_estimate = FALSE,
  correct_catch = FALSE
)

\S4method{SAM2FLStock}{sam,FLStock}(
  object,
  stk,
  uncertainty = FALSE,
  conf_level = 95,
  catch_estimate = FALSE,
  correct_catch = FALSE
)

\S4method{SAM2FLStock}{sam_list,FLStock}(
  object,
  stk,
  uncertainty = FALSE,
  conf_level = 95,
  catch_estimate = FALSE,
  correct_catch = FALSE
)
}
\arguments{
\item{object}{Object of class \code{sam} with the results from a
SAM stock assessment run. Alternatively, object of class \code{sam_list},
i.e. a list of \code{sam} objects.}

\item{stk}{Optional. Object of class \linkS4class{FLStock}, to which the
assessment results are added.}

\item{uncertainty}{If set to \code{TRUE}, the estimated uncertainty from
SAM will be added as attribute.}

\item{conf_level}{Confidence level used when uncertainty is returned.
Defaults to 95 (percent).}

\item{catch_estimate}{If set to \code{TRUE}, the catch estimated by SAM will
be returned instead of the model input.}

\item{correct_catch}{If set to \code{TRUE}, returned catch are correct by
catch multiplier estimated by SAM.}
}
\value{
An object of class \code{FLStock}.
}
\description{
This function takes the output from running the SAM stockassessment and
converts them into an \code{FLStock} object.
}
\details{
If an \code{FLStock} is provided as \code{stk} argument, then only the
following slots are updated: \code{stock}, \code{stock.n} & \code{harvest},
and, if requested by \code{catch_estimate = TRUE}, \code{catch},
\code{catch.n}, \code{landings}, \code{landings.n}, \code{discards},
\code{discards.n} with model estimates. If the dimensions
(years, iterations) differ between the SAM results and the provided stock
template, the returned \code{FLStock} is expanded.
}
\examples{
# fit SAM to North Sea cod
fit <- FLR_SAM(stk = cod4_stk, idx = cod4_idx, conf = cod4_conf_sam)

# coerce the output into FLStock
stk <- SAM2FLStock(fit)

# get catch estimates from model
stk <- SAM2FLStock(fit, catch_estimate = TRUE)

}
