% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{SAM_uncertainty}
\alias{SAM_uncertainty}
\alias{SAM_uncertainty,sam-method}
\title{Create replicates/iterations of SAM model fit based on variance-covariance
matrix}
\usage{
SAM_uncertainty(
  fit,
  n = 1000,
  print_screen = FALSE,
  seed = NULL,
  idx_cov = TRUE,
  catch_est = TRUE
)

\S4method{SAM_uncertainty}{sam}(
  fit,
  n = 1000,
  print_screen = FALSE,
  seed = NULL,
  idx_cov = TRUE,
  catch_est = TRUE
)
}
\arguments{
\item{fit}{A SAM model fit object of class \code{sam}.}

\item{n}{Number of replicates}

\item{print_screen}{If set to \code{TRUE}, print output of \code{TMB::sdreport} to screen.}

\item{idx_cov}{If set to \code{TRUE}, return covariance of survey index/indices.}

\item{catch_est}{If set to \code{TRUE}, return catch estimates from SAM.}
}
\value{
A list of FLQuants with the elements: stock.n, harvest, catch.n, catch_sd, survey_catchability, survey_sd, survey_cov, proc_error.
}
\description{
This function use the uncertainty estimated by SAM to create
replicates/iterations of assessment results. The function uses the
variance-covariance matrix to quantify uncertainty.
}
\details{
The returned objects are \code{FLQuant}s where the iteration dimension contains the
replicates. Each replicate is internally consistent, e.g. the fishing
mortality matches the stock numbers of the same replicate.

The following metrics are returned:
\itemize{
  \item{\code{stock.n}} Stock numbers at age for all years, \code{FLQuant}
  \item{\code{harvest}} Fishing mortalities at age for all years,
    \code{FLQuant}
  \item{\code{catch.n}} Estimates of catch numbers at age for all years. This
    differs from the assessment input values. If the SAM model fit contains
    catch multipliers, the values returned here are corrected for this.
    Class \code{FLQuant}.
  \item{\code{catch_sd}} Standard deviation of the catch numbers at age,
    time invariant, \code{FLQuant}
  \item{\code{survey_catchability}} Catchability at age for all years for all
    survey indices, list of \code{FLQuant}s
  \item{\code{survey_sd}} Standard deviation of all surveys at age, time
    invariant, list of \code{FLQuants},
  \item{\code{survey_cov}} Covariance matrices of survey ages, one for each
    survey. Return object is a list of lists, the first level corresponds to
    the replicates, the second level to the surveys. If no covariance
    between ages is assumed in the SAM model, the diagonal in the covariance
    matrices is simply the square root of the standard deviation
    (in \code{survey_sd})
  \item{\code{proc_error}} Standard deviation of the stock numbers at age,
    time invariant, class \code{FLQuant}. This corresponds to the survival
    process error assumed/estimated, i.e. quantifies how much the actual
    stock numbers at age deviate from the deterministic catch equation.
}
}
\examples{
### fit SAM to North Sea cod
fit <- FLR_SAM(stk = cod4_stk, idx = cod4_idx, conf = cod4_conf_sam)

### create 2 replicates
reps <- SAM_uncertainty(fit, n = 2)
}
