% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{whg4_stk}
\alias{whg4_stk}
\title{North Sea whiting}
\format{An object of class \code{FLStock}}
\source{
\url{https://www.stockassessment.org}
}
\usage{
whg4_stk
}
\description{
Assessment input data for North Sea whiting as used by ICES WGNSSK 2018.
The time series covers the years 1978-2018, with catches up to and
including 2017.
}
\keyword{datasets}
